Set build_wait_timeout to zero
This is required to have the service removed in a timely fashion when
the build is finished. The default value for `build_wait_timeout` is
10 minutes.

More specifically, `AbstractLatentWorker.buildFinished` schedules a
soft disconnect using `reactor.callLater` with the specified
timeout. The soft disconnect invokes insubstantiate after closing the
connection, which in turn invokes our `stop_instance`.

Here is the code path leading to service removal when a build is
finished:

  Build.allStepsDone
  Build.buildFinished
  AbstractWorkerForBuilder.buildFinished
  AbstractLatentWorker.buildFinished
  AbstractLatentWorker._soft_disconnect (via reactor.callLater)
  AbstractLatentWorker.insubstantiate
  DockerSwarmLatentWorker.stop_instance
  docker.models.Service.remove (via deferToThread)

Before invoking `insubstantiate`, `_soft_disconnect` forcibly
disconnects the worker using `AbstractWorker.disconnect`.

Here is the sequence of log messages with associated functions:

    " %s: build finished" (Build.buildFinished)
    "disconnecting old worker %s now" (AbstractWorker.disconnect)
    "waiting for worker to finish disconnecting" (AbstractWorker._disconnect)
    "insubstantiating worker {}" (AbstractLatentWorker.insubstantiate)

Some issues remain unresolved, though. As mentioned in the
`AbstractWorker.disconnect` docstring,

    If the worker is still alive, they will probably try to reconnect
    again in a moment.

This leads to the following log message, as well as a backtrace with
an identical exception message:

    Worker worker received connection while not trying to
    substantiate.  Disconnecting.

Around the same time, we also observe a sequence of log messages:

    20:49:34+0000 [-] Timing out client: IPv4Address(type='TCP', host='10.255.0.2', port=54388)
    20:49:50+0000 [-] Timing out client: IPv4Address(type='TCP', host='10.255.0.2', port=54389)
    20:49:50+0000 [-] Timing out client: IPv4Address(type='TCP', host='10.255.0.2', port=54390)
    20:49:50+0000 [-] Timing out client: IPv4Address(type='TCP', host='10.255.0.2', port=54386)
    20:49:50+0000 [-] Timing out client: IPv4Address(type='TCP', host='10.255.0.2', port=54391)
    20:49:50+0000 [-] Timing out client: IPv4Address(type='TCP', host='10.255.0.2', port=54387)

These appear to origin in the web frontend, however (see
`twisted.web.http.HTTPChannel.timeoutConnection`). Note that the IP
address recorded for the worker during its reconnect is not even on
the same network:

    worker 'worker' attaching from IPv4Address(type='TCP', host='10.0.24.6', port=38290)